pyproject.toml
[project]
name = "entropy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "boto3>=1.42.34",
    "docker>=7.1.0",
    "fastapi>=0.128.0",
    "fastmcp>=2.14.3",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.40.0",
]

[tool.uv.workspace]
members = [
    "mcp-servers/data-refinery",
    "mcp-servers/visualizer",
]

[dependency-groups]
dev = []
